# build/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UserIndex": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					}
				},
				"required": [
					"firstName",
					"lastName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProductIndex": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"$ref": "#/components/schemas/UserIndex"
					}
				},
				"required": [
					"id",
					"title",
					"description",
					"userId",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Product": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"userId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"title",
					"description",
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateProductRequest": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateProductRequest": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorMessage": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Session": {
				"properties": {
					"errors": {
						"items": {
							"$ref": "#/components/schemas/ErrorMessage"
						},
						"type": "array"
					},
					"token": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"errors",
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginUserRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"email",
					"firstName",
					"lastName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterUserRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"firstName",
					"lastName",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "training-node",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/products": {
			"get": {
				"operationId": "GetAllProducts",
				"responses": {
					"200": {
						"description": "200 - List of products",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ProductIndex"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get Products.",
				"summary": "Get all products.",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/products/create": {
			"post": {
				"operationId": "CreateProduct",
				"responses": {
					"200": {
						"description": "200 - Product",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Product"
								}
							}
						}
					}
				},
				"description": "Create Product.",
				"summary": "Create new product in database",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateProductRequest"
							}
						}
					}
				}
			}
		},
		"/products/{productId}": {
			"delete": {
				"operationId": "DeleteProduct",
				"responses": {
					"200": {
						"description": "200 - Product",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Product"
								}
							}
						}
					}
				},
				"description": "Delete Product.",
				"summary": "Delete product from database",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateProduct",
				"responses": {
					"200": {
						"description": "200 - Product",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Product"
								}
							}
						}
					}
				},
				"description": "Update Product.",
				"summary": "Update product from database",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateProductRequest"
							}
						}
					}
				}
			}
		},
		"/session/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "200 - Token",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Session"
								}
							}
						}
					}
				},
				"description": "Login User.",
				"summary": "Login user in app.",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginUserRequest"
							}
						}
					}
				}
			}
		},
		"/users": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "200 - User",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/User"
										},
										{
											"items": {
												"$ref": "#/components/schemas/ErrorMessage"
											},
											"type": "array"
										}
									]
								}
							}
						}
					}
				},
				"description": "Register User.",
				"summary": "Register new user in database.",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterUserRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "200 - List of users",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UserIndex"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Index - Get all users.",
				"summary": "Get a list of all users.",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}